Replace the asset-route introspection mock with a real database
The inherited-asset GET endpoints and the package-registry tarball
endpoints were tested through a mock that interpreted the tenant,
principal, and asset queries by walking drizzle's internal queryChunks
structure, duplicating the fragile introspection the resolver suites
carried. Those route tests now build the app against a migrated
postgres schema with real tenant, principal, and asset rows, so the
middleware and the asset resolver run their real queries.

The predicate-ignoring stub that backs the create and validation tests
stays: it returns fixed rows without inspecting any query, so it is
not exposed to drizzle's internal representation and needs no database.
